Separated integration tests to new make target (requires docker-compose), added small (test) supervised model.
