+++ .gitignore
@@ -7,5 +7,3 @@
 
 cover/
 tests/server/**/.cache/
-
-supervised_models/*.bin
