+++ Dockerfile
@@ -3,7 +3,7 @@ FROM python:3.6
 COPY . /dp-conceptual-search
 WORKDIR /dp-conceptual-search
 
-RUN make all clean
+RUN make build clean
 
 ENTRYPOINT ["python"]
 CMD ["manager.py"]
