+++ Dockerfile.base
@@ -1,0 +1,6 @@
+FROM python:3.6
+
+COPY . /dp-conceptual-search
+WORKDIR /dp-conceptual-search
+
+RUN make build clean
