--- Dockerfile.test
@@ -1,7 +1,0 @@
-FROM python:3.6
-
-COPY . /dp-conceptual-search
-WORKDIR /dp-conceptual-search
-
-RUN pip install -r requirements_test.txt
-RUN make build requirements_test && nosetests -v -s . && make clean
