+++ Makefile
@@ -7,15 +7,18 @@ build: requirements fastText
 requirements:
 	pip install -r requirements.txt
 
+test_requirements:
+	pip install -r requirements_test.txt
+
 fastText:
 	pip install Cython==0.27.3 pybind11==2.2.3
 	cd lib/fastText && python setup.py install
 
-test:
-	pip install -r requirements_test.txt
-	docker-compose up -d
+test: test_requirements
 	python manager.py test
-	docker-compose down
+
+integration-test: test_requirements
+	nosetests -s -v tests.integration
 
 pep8:
 	autopep8 --in-place --aggressive --aggressive -r ./
