+++ ci/scripts/unit.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -eux
 
 pushd dp-conceptual-search
-  make build test && make clean
+  make build test clean
 popd
