+++ config_testing.py
@@ -4,7 +4,7 @@ TESTING = True
 ENABLE_PROMETHEUS_METRICS = False
 
 MONGO_SEARCH_DATABASE = 'test'
-MONGO_BIND_ADDR = 'mongodb://0.0.0.0:27017'
+MONGO_BIND_ADDR = 'mongodb://mongo:27017'
 
 MOTOR_URI = "{bind_addr}/{db}".format(
     bind_addr=MONGO_BIND_ADDR,
