+++ requirements_test.txt
@@ -1,2 +1,3 @@
 nose==1.3.7
+nose-exclude==0.5.0
 coverage==4.5.1
